requirements.txt
bayesian_changepoint_detection==0.2.dev1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
fonttools==4.62.1
frozendict==2.4.7
hmmlearn==0.3.3
idna==3.11
joblib==1.5.3
kiwisolver==1.5.0
matplotlib==3.10.8
multitasking==0.0.12
numpy==2.4.4
nvidia-nccl-cu12==2.29.7
packaging==26.0
pandas==3.0.2
peewee==4.0.4
pillow==12.2.0
platformdirs==4.9.4
protobuf==7.34.1
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
soupsieve==2.8.3
threadpoolctl==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
xgboost==3.2.0
yfinance==1.2.0
